Move request utilities to helper module
